split

+++ ft_bzero.c
@@ -2,7 +2,7 @@
 #include <unistd.h>
 #include <string.h>
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	int		i;
 	char	*tab;
